2.6.0 merge to master (#606)
* 修改chainid强转short的bug

* 修复cmd依赖问题

* broad cross chain info change tx remove main chain

* cross chain info change handler update

* chainManager cross asset record update

* 单元测试

* 修改忽略文件配置

* 修改版本号 (#600)

* 修改版本号 (#601)

* 2.6.0 修改协议版本号 (#602)

* 修改版本号

* protocal config update


* 命令行支持动态执行功能

* update  getblock  api

* update  getblock  api

* nuls-api增加链内转平行链资产的离线交易组装接口

* 修改平行链资产小数位错误的问题。

* 修复链管理在注销资产时验证资产占比时小数位错误的问题

* 修复增加跨链资产时小数位错误的问题

+++ module/nuls-consensus/nuls-consensus-poc/src/main/java/io/nuls/poc/utils/validator/TxValidator.java
@@ -251,7 +251,7 @@ public class TxValidator {
      * @return boolean
      */
     private boolean createAgentBasicValid(Chain chain, Transaction tx, Agent agent) throws NulsException {
-        if (!AddressTool.validNormalAddress(agent.getPackingAddress(), (short) chain.getConfig().getChainId())) {
+        if (!AddressTool.validNormalAddress(agent.getPackingAddress(), chain.getConfig().getChainId())) {
             throw new NulsException(ConsensusErrorCode.ADDRESS_ERROR);
         }
         if (Arrays.equals(agent.getAgentAddress(), agent.getPackingAddress())) {
